+++ js/jquery.terminal.js
@@ -31,7 +31,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Fri, 23 Jun 2017 07:02:47 +0000
+ * Date: Sat, 24 Jun 2017 13:14:15 +0000
  */
 
 /* TODO:
@@ -2764,7 +2764,9 @@
                 // remove quotes if before are even number of slashes
                 // we don't remove slases becuase they are handled by JSON.parse
                 string = string.replace(/((^|[^\\])(?:\\\\)*)['"]/g, '$1');
+                string = string.replace(/([^\\])['"]$/g, '$1');
                 // use build in function to parse rest of escaped characters
+                console.log(string);
                 return JSON.parse('"' + string + '"');
             }
             if (strict === false) {