+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 * add option to cmd::disable to not blur so it don't hide android keyboard
   on terminal::pause(true)
+* don't enable terminal on init on Android
 
 # 1.4.3
 * don't execute keypress callback when terminal is disabled (reported by @artursOs)