+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   on terminal::pause(true)
 * don't enable terminal on init on Android
 * fix next key after CTRL+V that was triggering input event (reported by @artursOs)
+* fix parsing strings
 
 # 1.4.3
 * don't execute keypress callback when terminal is disabled (reported by @artursOs)