+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * don't enable terminal on init on Android
 * fix next key after CTRL+V that was triggering input event (reported by @artursOs)
 * fix parsing strings
+* don't hide virtual keyboard on Android when called pause()
 
 # 1.4.3
 * don't execute keypress callback when terminal is disabled (reported by @artursOs)