+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix parsing strings
 * don't hide virtual keyboard on Android when called pause()
 * fix input on Firefox with google keyboard (reported by Filip Wieland)
+* disable terminal on pause event in cordova
 
 # 1.4.3
 * don't execute keypress callback when terminal is disabled (reported by @artursOs)