+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 # next
-* add option to cmd::disable to not blur so it don't hide android keyboard
-  on terminal::pause(true)
+* add option to cmd::disable to not blur so it don't hide android keyboard on pause
 * don't enable terminal on init on Android
 * fix next key after CTRL+V that was triggering input event (reported by @artursOs)
 * fix parsing strings
 * don't hide virtual keyboard on Android when called pause()
 * fix input on Firefox with google keyboard (reported by Filip Wieland)
-* disable terminal on pause event in cordova
+* disable terminal on resume event in cordova
 
 # 1.4.3
 * don't execute keypress callback when terminal is disabled (reported by @artursOs)